pyproject.toml
[tool.poetry]
name = "project_icon_maker"
version = "0.1.0"
description = ""
authors = ["enrique <enriquesoria@cecotec.es>"]

[tool.poetry.dependencies]
python = "^3.6"
Pillow = "^8.2.0"
emoji = "^1.2.0"
pycairo = "^1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
